+++ R/qenv-constructor.R
@@ -1,24 +1,52 @@
 #' Initialize `qenv` object
 #'
-#' Initialize `qenv` object with `code` and `env`. In order to have `qenv` reproducible
-#' one needs to initialize with `env` which can be reproduced by the `code`. Alternatively, one
-#' can create an empty `qenv` and evaluate the expressions in this object using `eval_code`.
-#' @name new_qenv
+#' Create an empty `qenv` object.
 #'
-#' @param code (`character(1)` or `language`) code to evaluate. Accepts and stores comments also.
-#' @param env (`environment`) Environment being a result of the `code` evaluation.
+#' `qenv()` instantiates a `qenv` with an empty environment.
+#' Any changes must be made by evaluating code in it with `eval_code` or `within`, thereby ensuring reproducibility.
+#'
+#' `new_qenv()` (deprecated and not recommended)
+#' can instantiate a `qenv` object with data in the environment and code registered.
+#'
+#' @name qenv
 #'
 #' @examples
-#' new_qenv(env = list2env(list(a = 1)), code = quote(a <- 1))
-#' new_qenv(env = list2env(list(a = 1)), code = parse(text = "a <- 1", keep.source = TRUE))
-#' new_qenv(env = list2env(list(a = 1)), code = "a <- 1")
+#' # create empty qenv
+#' qenv()
 #'
 #' @return `qenv` object.
 #'
 #' @export
-setGeneric("new_qenv", function(env = new.env(parent = parent.env(.GlobalEnv)), code = character()) standardGeneric("new_qenv")) # nolint
+qenv <- function() {
+  q_env <- new.env(parent = parent.env(.GlobalEnv))
+  lockEnvironment(q_env, bindings = TRUE)
+  methods::new("qenv", env = q_env)
+}
 
-#' @rdname new_qenv
+
+#' @param code `r badge("deprecated")`
+#'  (`character(1)` or `language`) code to evaluate. Accepts and stores comments also.
+#' @param env `r badge("deprecated")` (`environment`)
+#'  Environment being a result of the `code` evaluation.
+#'
+#' @examples
+#' # create qenv with data and code (deprecated)
+#' new_qenv(env = list2env(list(a = 1)), code = quote(a <- 1))
+#' new_qenv(env = list2env(list(a = 1)), code = parse(text = "a <- 1", keep.source = TRUE))
+#' new_qenv(env = list2env(list(a = 1)), code = "a <- 1")
+#'
+#' @rdname qenv
+#' @aliases new_qenv,environment,expression-method
+#' @aliases new_qenv,environment,character-method
+#' @aliases new_qenv,environment,language-method
+#' @aliases new_qenv,environment,missing-method
+#' @aliases new_qenv,missing,missing-method
+#' @export
+setGeneric("new_qenv", function(env = new.env(parent = parent.env(.GlobalEnv)), code = character()) {
+  lifecycle::deprecate_warn(when = " 0.4.2", what = "new_qenv()", with = "qenv()", always = TRUE)
+  standardGeneric("new_qenv")
+})
+
 #' @export
 setMethod(
   "new_qenv",
@@ -28,7 +56,6 @@ setMethod(
   }
 )
 
-#' @rdname new_qenv
 #' @export
 setMethod(
   "new_qenv",
@@ -45,7 +72,6 @@ setMethod(
   }
 )
 
-#' @rdname new_qenv
 #' @export
 setMethod(
   "new_qenv",
@@ -55,7 +81,6 @@ setMethod(
   }
 )
 
-#' @rdname new_qenv
 #' @export
 setMethod(
   "new_qenv",