+++ tests/testthat/test-qenv-replace_code.R
@@ -1,4 +1,4 @@
-q <- new_qenv()
+q <- qenv()
 qq <- within(q, i <- iris)
 
 # argument checks ----