+++ tests/testthat/test-qenv_get_var.R
@@ -1,5 +1,5 @@
 testthat::test_that("get_var and `[[`return error if object is qenv.error", {
-  q <- eval_code(new_qenv(), quote(x <- 1))
+  q <- eval_code(qenv(), quote(x <- 1))
   q <- eval_code(q, quote(y <- w * x))
 
   testthat::expect_error(get_var(q, "x"), "when evaluating qenv code")
@@ -7,7 +7,7 @@ testthat::test_that("get_var and `[[`return error if object is qenv.error", {
 })
 
 testthat::test_that("get_var and `[[` return object from qenv environment", {
-  q <- eval_code(new_qenv(), quote(x <- 1))
+  q <- eval_code(qenv(), quote(x <- 1))
   q <- eval_code(q, quote(y <- 5 * x))
 
   testthat::expect_equal(get_var(q, "y"), 5)
@@ -15,7 +15,7 @@ testthat::test_that("get_var and `[[` return object from qenv environment", {
 })
 
 testthat::test_that("get_var and `[[` return NULL if object not in qenv environment", {
-  q <- eval_code(new_qenv(), quote(x <- 1))
+  q <- eval_code(qenv(), quote(x <- 1))
   q <- eval_code(q, quote(y <- 5 * x))
 
   testthat::expect_null(get_var(q, "z"))